// apps/pokeverse/src/helpers/types.ts
export interface PokemonSprite {
    front_default: string;
    front_shiny: string;
    back_default?: string;
    back_shiny?: string;
    other: {
        'official-artwork': {
            front_default: string;
            front_shiny: string;
        };
        home: {
            front_default: string;
            front_shiny: string;
        };
        showdown: {
            front_default: string;
            back_default: string;
        };
    };
}
// ...
export interface Move {
    name: string;
    url: string;
}

export interface Ability {
    name: string;
    isHidden: boolean;
}

export interface Stats {
    name: string;
    value: number;
}

export interface AltForm {
    name: string;
    id: number;
}

export interface EvolutionDetails {
    name: string;
    id: number;
    evolvesTo: EvolutionDetails[];
}

export interface Pokemon {
    id: number;
    name: string;
    baseExp: number;
    height: string;
    weight: string;
    types: string[];
    abilities: Ability[];
    sprites: PokemonSprite;
    moves: Move[];
    stats: Stats[];

    baseHappiness: number;
    captureRate: string;
    flavorTexts: string[];
    generation: number;
    genre: string;
    habitat: string;
    romanjiName?: string;
    varieties: AltForm[];

    evolutionChain: EvolutionDetails;

    speciesId?: number;
    evoChainId?: number;
}